LRAC-11218 change getSessionsPerVisitor from long to double

+++ modules/dxp/apps/osb/osb-asah/osb-asah-common/src/main/java/com/liferay/osb/asah/common/model/VisitorBehaviorMetric.java
@@ -52,7 +52,7 @@ public abstract class VisitorBehaviorMetric {
 		return _sessions.longValue();
 	}
 
-	public long getSessionsPerVisitor() {
+	public double getSessionsPerVisitor() {
 		if ((_sessions != null) && (_visitors != null) &&
 			!_visitors.equals(BigDecimal.ZERO)) {
